requirements.txt
futures==3.2.0
BeautifulSoup==3.2.1
pexpect==4.5.0
psutil==5.4.1
pycrypto==2.6.1
pyserial==3.4
netifaces==0.10.7
numpy==1.15.2
PyYAML==3.13
PyDispatcher==2.0.5
usbinfo==1.1.0
dpkt==1.9.1
dispatcher==1.0
jsonfield==2.0.2
urllib3==1.23
ipaddress==1.0.22
